// frontend/src/app/components/project/components/resources/parser-config/config-type.enum.ts
export enum GPStandardConfigEnum {
    SSMLStandard = 'SSML',
    SRGSStandard = 'SRGS',
}

export enum GPConfigFileTypeEnum {
    COMMAND = 'Command',
    PROMPT = 'Prompt',
}

export enum GPConfigurationErrorMessageEnum {
    SSMLEditor = 'SSMLEditor',
    SSMLTranslator = 'SSMLTranslator',
    SRGSEditor = 'SRGSEditor',
    SRGSTranslator = 'SRGSTranslator',
}

export enum GPProjectRoleEnum {
    Editor = 'Editor',
    Translator = 'Translator',
}

export enum GPStandardFileNameEnum {
    fileName = 'Standard Configuration',
}